Check user before edit/delete

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
@@ -93,13 +94,21 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session,
+                            RedirectAttributes redirectAttributes) {
         Image image = imageService.getImage(imageId);
 
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        // if currently logged in user is the image.getUser then
+        User currentUser = (User) session.getAttribute("loggeduser");
+        if(image.getUser().equals(currentUser)) {
+            //String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            //model.addAttribute("tags", tags);
+            return "images/edit";
+        }else{
+            redirectAttributes.addFlashAttribute("editError", true);
+            return "redirect:/images/" + image.getId();
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -141,9 +150,20 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,Model model,HttpSession session,
+                                    RedirectAttributes redirectAttributes) {
+
+        Image image = imageService.getImage(imageId);
+
+        // if currently logged in user is the image.getUser then
+        User currentUser = (User) session.getAttribute("loggeduser");
+        if(image.getUser().equals(currentUser)) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }else{
+            redirectAttributes.addFlashAttribute("deleteError", true);
+            return "redirect:/images/" + image.getId();
+        }
     }
